// github/types/paths/repos/{owner}/{repo}/pulls/{pull_number}.types.ts
export type pullsGet = (
  $: OmitValueWhenNever<{
    query: never;
    querystring: never;
    path: pullsGet_Path;
    headers: never;
    cookie: never;
    body: never;
    context: Context;
    response: ResponseBuilderFactory<{
      200: {
        headers: never;
        requiredHeaders: never;
        content: {
          "application/json": {
            schema: pull_request;
          };
        };
        examples: {
          default: unknown;
        };
      };
      304: not_modified;
      404: not_found;
      406: unacceptable;
      500: internal_error;
      503: service_unavailable;
    }>;
    x: WideOperationArgument;
    proxy: (url: string) => COUNTERFACT_RESPONSE;
    user: never;
    delay: (milliseconds: number, maxMilliseconds?: number) => Promise<void>;
    version: never;
  }>,
) => MaybePromise<COUNTERFACT_RESPONSE>;
// ...
export type pullsUpdate = (
  $: OmitValueWhenNever<{
    query: never;
    querystring: never;
    path: pullsUpdate_Path;
    headers: never;
    cookie: never;
    body: {
      /**
       * The title of the pull request.
       */
      title?: string;
      /**
       * The contents of the pull request.
       */
      body?: string;
      /**
       * State of this Pull Request. Either `open` or `closed`.
       */
      state?: "open" | "closed";
      /**
       * The name of the branch you want your changes pulled into. This should be an existing branch on the current repository. You cannot update the base branch on a pull request to point to another repository.
       */
      base?: string;
      /**
       * Indicates whether [maintainers can modify](https://docs.github.com/articles/allowing-changes-to-a-pull-request-branch-created-from-a-fork/) the pull request.
       */
      maintainer_can_modify?: boolean;
    };
    context: Context;
    response: ResponseBuilderFactory<{
      200: {
        headers: never;
        requiredHeaders: never;
        content: {
          "application/json": {
            schema: pull_request;
          };
        };
        examples: {
          default: unknown;
        };
      };
      403: forbidden;
      422: validation_failed;
    }>;
    x: WideOperationArgument;
    proxy: (url: string) => COUNTERFACT_RESPONSE;
    user: never;
    delay: (milliseconds: number, maxMilliseconds?: number) => Promise<void>;
    version: never;
  }>,
) => MaybePromise<COUNTERFACT_RESPONSE>;

export type pullsGet_Path = {
  owner: string;
  repo: string;
  pull_number: number;
};

export type pullsUpdate_Path = {
  owner: string;
  repo: string;
  pull_number: number;
};